requirements.txt
requests==2.21.0
PyYAML==3.13
future-fstrings==0.4.5
futures==3.1.1
six==1.12.0
google-api-python-client==1.7.8
oauth2client==4.1.3